FIX: Send prox events to widget on hardware prox
The widget focus is not cleared when a tool goes out of prox. Not only does this
prevent the focused widget from ever seeing a hardware out-prox event, it also
prevents it from seeing a hardware in-prox event if it was the same widget the
tool left prox over. This patch fixes the issue by sending the out-prox event
on hardware events, and ensures the widget focus is cleared so that the in-prox
can be seen (via tablet motion -- remember that in prox itself has no coordinates!)